+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.0 branch
 ### 1.0.2-beta3
+* command writes number of added files;
 * added `--debug` options for `FilesBackupCommand`.
 
 ### 1.0.1-beta2